Separate logging components
Break out logging components into individual profiles so that
logging can be a pluggable component, as well as supporting
roles that separate the data part of log storage from the
indexing and filtering part of log aggregation as well as the
dashboard.

Elasticsearch is moved to the database section as it may be consumed
by a non-logging application.